Move crypto errors in to the main error submodule of sign.
