+++ src/sign/crypto/common.rs
@@ -10,7 +10,7 @@ use std::sync::Arc;
 use ::ring::rand::SystemRandom;
 
 use crate::base::iana::SecAlg;
-use crate::sign::error::SignError;
+use crate::sign::error::{FromBytesError, GenerateError, SignError};
 use crate::sign::{SecretKeyBytes, SignRaw};
 use crate::validate::{PublicKeyBytes, Signature};
 
@@ -195,113 +195,6 @@ pub fn generate(
     // Otherwise fail.
     #[allow(unreachable_code)]
     Err(GenerateError::UnsupportedAlgorithm)
-}
-
-//============ Error Types ===================================================
-
-//----------- FromBytesError -----------------------------------------------
-
-/// An error in importing a key pair from bytes.
-#[derive(Clone, Debug)]
-pub enum FromBytesError {
-    /// The requested algorithm was not supported.
-    UnsupportedAlgorithm,
-
-    /// The key's parameters were invalid.
-    InvalidKey,
-
-    /// The implementation does not allow such weak keys.
-    WeakKey,
-
-    /// An implementation failure occurred.
-    ///
-    /// This includes memory allocation failures.
-    Implementation,
-}
-
-//--- Conversions
-
-#[cfg(feature = "ring")]
-impl From<ring::FromBytesError> for FromBytesError {
-    fn from(value: ring::FromBytesError) -> Self {
-        match value {
-            ring::FromBytesError::UnsupportedAlgorithm => {
-                Self::UnsupportedAlgorithm
-            }
-            ring::FromBytesError::InvalidKey => Self::InvalidKey,
-            ring::FromBytesError::WeakKey => Self::WeakKey,
-        }
-    }
-}
-
-#[cfg(feature = "openssl")]
-impl From<openssl::FromBytesError> for FromBytesError {
-    fn from(value: openssl::FromBytesError) -> Self {
-        match value {
-            openssl::FromBytesError::UnsupportedAlgorithm => {
-                Self::UnsupportedAlgorithm
-            }
-            openssl::FromBytesError::InvalidKey => Self::InvalidKey,
-            openssl::FromBytesError::Implementation => Self::Implementation,
-        }
-    }
-}
-
-//--- Formatting
-
-impl fmt::Display for FromBytesError {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str(match self {
-            Self::UnsupportedAlgorithm => "algorithm not supported",
-            Self::InvalidKey => "malformed or insecure private key",
-            Self::WeakKey => "key too weak to be supported",
-            Self::Implementation => "an internal error occurred",
-        })
-    }
-}
-
-//--- Error
-
-impl std::error::Error for FromBytesError {}
-
-//----------- GenerateError --------------------------------------------------
-
-/// An error in generating a key pair.
-#[derive(Clone, Debug)]
-pub enum GenerateError {
-    /// The requested algorithm was not supported.
-    UnsupportedAlgorithm,
-
-    /// An implementation failure occurred.
-    ///
-    /// This includes memory allocation failures.
-    Implementation,
-}
-
-//--- Conversion
-
-#[cfg(feature = "ring")]
-impl From<ring::GenerateError> for GenerateError {
-    fn from(value: ring::GenerateError) -> Self {
-        match value {
-            ring::GenerateError::UnsupportedAlgorithm => {
-                Self::UnsupportedAlgorithm
-            }
-            ring::GenerateError::Implementation => Self::Implementation,
-        }
-    }
-}
-
-#[cfg(feature = "openssl")]
-impl From<openssl::GenerateError> for GenerateError {
-    fn from(value: openssl::GenerateError) -> Self {
-        match value {
-            openssl::GenerateError::UnsupportedAlgorithm => {
-                Self::UnsupportedAlgorithm
-            }
-            openssl::GenerateError::Implementation => Self::Implementation,
-        }
-    }
 }
 
 //--- Formatting
