+++ src/sign/crypto/mod.rs
@@ -1,3 +1,4 @@
+//! Cryptographic backends.
 pub mod common;
 pub mod openssl;
 pub mod ring;
